requirements-prod.txt
gunicorn==20.1.0
django-anymail[mailgun]==9.1
